+++ src/git.ts
@@ -1,5 +1,6 @@
 import { exec } from '@actions/exec';
 import { context, getOctokit } from '@actions/github';
+import * as fs from 'node:fs';
 import { readPackageJSON, resolvePackageJSON, writePackageJSON } from 'pkg-types';
 import { COMMIT_TEMPLATES, ERROR_MESSAGES, GIT_USER_CONFIG } from './constants';
 import core, { logger } from './core';
@@ -62,8 +63,10 @@ export async function configureGitUser(): Promise<void> {
  */
 export async function hasFileChanges(filepath: string): Promise<boolean> {
   try {
-    // 检查文件是否存在
-    await exec('test', ['-f', filepath]);
+    // 检查文件是否存在 - 使用 Node.js fs 模块更可靠
+    if (!fs.existsSync(filepath)) {
+      return false;
+    }
 
     // 检查是否有变化
     const statusOutput = await execGitWithOutput(['status', '--porcelain', filepath]);
@@ -160,222 +163,6 @@ async function safePushWithRetry(targetBranch: SupportedBranch, version: string,
   }
 }
 
-// ==================== CHANGELOG 操作 ====================
-
-/**
- * 基于PR信息生成CHANGELOG条目
- */
-async function generateChangelogFromPR(pr: PRData | null, version: string): Promise<string> {
-  if (!pr) {
-    return `### Changes\n- Version ${version} release\n`;
-  }
-
-  // PR标签到CHANGELOG类型的映射
-  const labelToChangelogType: Record<string, string> = {
-    major: '💥 Breaking Changes',
-    minor: '✨ Features',
-    patch: '🐛 Bug Fixes',
-    enhancement: '⚡ Improvements',
-    performance: '🚀 Performance',
-    security: '🔒 Security',
-    documentation: '📚 Documentation',
-    dependencies: '⬆️ Dependencies',
-  };
-
-  // 从PR标签推断变更类型
-  let changeType = '📝 Changes';
-  if (pr.labels) {
-    for (let i = 0; i < pr.labels.length; i++) {
-      const label = pr.labels[i];
-      if (labelToChangelogType[label.name]) {
-        changeType = labelToChangelogType[label.name];
-        break;
-      }
-    }
-
-    // 如果没找到特定类型，基于版本标签推断
-    if (changeType === '📝 Changes') {
-      const versionLabels = pr.labels.map((l) => l.name);
-      if (versionLabels.includes('major')) changeType = '💥 Breaking Changes';
-      else if (versionLabels.includes('minor')) changeType = '✨ Features';
-      else if (versionLabels.includes('patch')) changeType = '🐛 Bug Fixes';
-    }
-  }
-
-  // 构建CHANGELOG条目
-  let changelogEntry = `### ${changeType}\n`;
-
-  // 添加PR标题和链接
-  const prUrl = pr.html_url;
-  const prTitle = pr.title || `PR #${pr.number}`;
-  changelogEntry += `- ${prTitle} ([#${pr.number}](${prUrl}))\n`;
-
-  // 如果PR有body，提取关键信息
-  if (pr.body && pr.body.trim()) {
-    const body = pr.body.trim();
-
-    // 查找特定的section（如 "### Changes", "## What's Changed" 等）
-    const sections = [
-      '### Changes',
-      '## Changes',
-      "### What's Changed",
-      "## What's Changed",
-      '### Summary',
-      '## Summary',
-    ];
-    for (let i = 0; i < sections.length; i++) {
-      const section = sections[i];
-      const sectionIndex = body.indexOf(section);
-      if (sectionIndex !== -1) {
-        const sectionContent = body.substring(sectionIndex + section.length);
-        const nextSectionIndex = sectionContent.search(/^##/m);
-        const content = nextSectionIndex !== -1 ? sectionContent.substring(0, nextSectionIndex) : sectionContent;
-
-        const cleanContent = content
-          .trim()
-          .split('\n')
-          .filter((line) => line.trim())
-          .slice(0, 5) // 最多5行
-          .map((line) => (line.startsWith('- ') ? `  ${line}` : `  - ${line}`))
-          .join('\n');
-
-        if (cleanContent) {
-          changelogEntry += `${cleanContent}\n`;
-          break;
-        }
-      }
-    }
-  }
-
-  return changelogEntry;
-}
-
-/**
- * 更新 CHANGELOG - 基于PR信息生成
- */
-export async function updateChangelog(pr: PRData | null = null, version: string = ''): Promise<void> {
-  // 检查是否启用CHANGELOG生成
-  const enableChangelog = core.getInput('enable-changelog')?.toLowerCase() !== 'false';
-  if (!enableChangelog) {
-    logger.info('CHANGELOG 生成已禁用，跳过');
-    return;
-  }
-
-  try {
-    logger.info('开始生成基于PR的 CHANGELOG...');
-
-    const currentDate = new Date().toISOString().split('T')[0];
-    const versionTag = version.startsWith('v') ? version : `v${version}`;
-
-    // 生成基于PR的CHANGELOG条目
-    const changelogEntry = await generateChangelogFromPR(pr, version);
-
-    const newEntry = `## [${versionTag}] - ${currentDate}
-
-${changelogEntry}
-`;
-
-    // 读取现有CHANGELOG内容
-    let existingContent = '';
-    try {
-      let stdout = '';
-      await exec('cat', ['CHANGELOG.md'], {
-        listeners: {
-          stdout: (data: Buffer) => {
-            stdout += data.toString();
-          },
-        },
-      });
-      existingContent = stdout;
-      logger.info('读取现有CHANGELOG内容');
-    } catch {
-      // 如果文件不存在，创建初始内容
-      existingContent = `# Changelog
-
-All notable changes to this project will be documented in this file.
-
-The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
-and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
-
-`;
-      logger.info('CHANGELOG.md 不存在，创建新文件');
-    }
-
-    // 插入新条目到第一个版本记录之前
-    const lines = existingContent.split('\n');
-    let insertIndex = lines.length;
-
-    // 查找第一个版本标题的位置
-    for (let i = 0; i < lines.length; i++) {
-      if (lines[i].match(/^## \[.*\]/)) {
-        insertIndex = i;
-        break;
-      }
-    }
-
-    // 插入新条目
-    const entryLines = newEntry.split('\n');
-    lines.splice(insertIndex, 0, ...entryLines);
-
-    // 写回文件
-    const newContent = lines.join('\n');
-    await exec('sh', ['-c', `cat > CHANGELOG.md << 'EOF'\n${newContent}\nEOF`]);
-
-    logger.info(`✅ CHANGELOG 已更新，添加版本 ${versionTag}`);
-
-    // 显示新增的内容预览
-    try {
-      let stdout = '';
-      await exec('head', ['-15', 'CHANGELOG.md'], {
-        listeners: {
-          stdout: (data: Buffer) => {
-            stdout += data.toString();
-          },
-        },
-      });
-      logger.info('📋 CHANGELOG 预览:');
-      logger.info(stdout);
-    } catch {
-      logger.info('无法显示CHANGELOG预览');
-    }
-  } catch (error) {
-    logger.warning(`基于PR的CHANGELOG生成失败: ${error}`);
-
-    // 如果失败，使用原来的conventional-changelog逻辑作为备用
-    await fallbackToConventionalChangelog();
-  }
-}
-
-/**
- * 备用方案：使用conventional-changelog
- */
-async function fallbackToConventionalChangelog(): Promise<void> {
-  try {
-    logger.info('使用conventional-changelog作为备用方案...');
-
-    // 检查是否已安装
-    try {
-      await exec('npx', ['conventional-changelog-cli', '--version']);
-    } catch {
-      await exec('npm', ['install', '-g', 'conventional-changelog-cli', 'conventional-changelog-conventionalcommits']);
-    }
-
-    await exec('npx', [
-      'conventional-changelog-cli',
-      '-p',
-      'conventionalcommits',
-      '-i',
-      'CHANGELOG.md',
-      '-s',
-      '-r',
-      '0',
-    ]);
-
-    logger.info('✅ 使用conventional-changelog生成完成');
-  } catch (error) {
-    logger.warning(`备用CHANGELOG生成也失败: ${error}`);
-  }
-}
 
 // ==================== 分支同步逻辑 ====================
 
@@ -663,170 +450,6 @@ export async function syncBranches(targetBranch: SupportedBranch, newVersion: st
   return results;
 }
 
-// ==================== NPM 发布功能 ====================
-
-/**
- * 检查是否启用npm发布
- */
-function isNpmPublishEnabled(): boolean {
-  const enablePublish = core.getInput('enable-npm-publish')?.toLowerCase();
-  return enablePublish === 'true';
-}
-
-/**
- * 获取npm发布配置
- */
-function getNpmPublishConfig() {
-  const registry = core.getInput('npm-registry') || 'https://registry.npmjs.org/';
-  const token = core.getInput('npm-token');
-  const tag = core.getInput('npm-tag') || 'latest';
-  const access = core.getInput('npm-access') || 'public';
-
-  return { registry, token, tag, access };
-}
-
-/**
- * 配置npm认证
- */
-async function configureNpmAuth(registry: string, token: string): Promise<void> {
-  try {
-    // 设置registry
-    await exec('npm', ['config', 'set', 'registry', registry]);
-    logger.info(`配置npm registry: ${registry}`);
-
-    // 设置认证token
-    if (token) {
-      const registryUrl = new URL(registry);
-      const authKey = `//${registryUrl.host}/:_authToken`;
-      await exec('npm', ['config', 'set', authKey, token]);
-      logger.info('配置npm认证token');
-    }
-  } catch (error) {
-    throw new ActionError(`配置npm认证失败: ${error}`, 'configureNpmAuth', error);
-  }
-}
-
-/**
- * 确定npm发布标签
- */
-function determineNpmTag(version: string, targetBranch: SupportedBranch, configTag: string): string {
-  // 如果用户指定了特定标签，使用用户指定的标签
-  if (configTag !== 'latest') {
-    return configTag;
-  }
-
-  // 根据分支和版本自动确定标签
-  if (targetBranch === 'main') {
-    // 主分支使用latest标签
-    return 'latest';
-  } else if (targetBranch === 'beta') {
-    // Beta分支使用beta标签
-    return 'beta';
-  } else if (targetBranch === 'alpha') {
-    // Alpha分支使用alpha标签
-    return 'alpha';
-  }
-
-  // 如果是预发布版本，根据prerelease标识确定标签
-  const cleanedVersion = cleanVersion(version);
-  const parsed = parseVersion(cleanedVersion);
-  if (parsed?.prerelease && parsed.prerelease.length > 0) {
-    const prereleaseId = parsed.prerelease[0] as string;
-    if (prereleaseId === 'alpha') return 'alpha';
-    if (prereleaseId === 'beta') return 'beta';
-  }
-
-  return 'latest';
-}
-
-/**
- * 执行npm发布
- */
-async function publishToNpm(
-  version: string,
-  targetBranch: SupportedBranch,
-  config: { registry: string; token: string; tag: string; access: string },
-): Promise<void> {
-  try {
-    // 确定发布标签
-    const publishTag = determineNpmTag(version, targetBranch, config.tag);
-
-    logger.info(`准备发布到npm: 版本=${version}, 标签=${publishTag}, 分支=${targetBranch}`);
-
-    // 配置npm认证
-    await configureNpmAuth(config.registry, config.token);
-
-    // 构建发布命令
-    const publishArgs = ['publish'];
-
-    // 添加访问权限
-    if (config.access) {
-      publishArgs.push('--access', config.access);
-    }
-
-    // 添加标签
-    publishArgs.push('--tag', publishTag);
-
-    // 执行发布
-    await exec('npm', publishArgs);
-
-    logger.info(`✅ 成功发布到npm: ${version} (标签: ${publishTag})`);
-
-    // 设置输出
-    core.setOutput('published-version', version);
-    core.setOutput('published-tag', publishTag);
-    core.setOutput('npm-registry', config.registry);
-  } catch (error) {
-    // 检查是否是版本已存在的错误
-    const errorMessage = String(error);
-    if (
-      errorMessage.includes('version already exists') ||
-      errorMessage.includes('You cannot publish over the previously published versions')
-    ) {
-      logger.warning(`版本 ${version} 已存在于npm registry，跳过发布`);
-      return;
-    }
-
-    throw new ActionError(`npm发布失败: ${error}`, 'publishToNpm', error);
-  }
-}
-
-/**
- * 处理npm发布逻辑 - 只对目标分支版本发布
- */
-export async function handleNpmPublish(version: string, targetBranch: SupportedBranch): Promise<void> {
-  if (!isNpmPublishEnabled()) {
-    logger.info('npm发布已禁用，跳过');
-    return;
-  }
-
-  try {
-    logger.info(`开始npm发布流程: 版本=${version}, 目标分支=${targetBranch}`);
-
-    const config = getNpmPublishConfig();
-
-    // 验证必需的配置
-    if (!config.token) {
-      throw new ActionError('npm-token未配置，无法发布到npm', 'handleNpmPublish');
-    }
-
-    // 只对目标分支的版本进行发布，不处理下游分支
-    await publishToNpm(version, targetBranch, config);
-
-    logger.info(`✅ ${targetBranch}分支版本 ${version} npm发布完成`);
-  } catch (error) {
-    // npm发布失败不应该中断整个流程
-    logger.error(`npm发布失败: ${error}`);
-    core.setOutput('npm-publish-failed', 'true');
-    core.setOutput('npm-publish-error', String(error));
-
-    // 如果用户要求严格模式，则抛出错误
-    const strictMode = core.getInput('npm-publish-strict')?.toLowerCase() === 'true';
-    if (strictMode) {
-      throw error;
-    }
-  }
-}
 
 // ==================== 版本更新和标签创建 ====================
 
@@ -851,19 +474,19 @@ export async function updateVersionAndCreateTag(
     await commitAndPushVersion(newVersion, targetBranch);
 
     // 🎯 在打tag后更新 CHANGELOG - 使用PR信息
+    const { updateChangelog, hasChangelogChanges, commitChangelog } = await import('./changelog');
     await updateChangelog(pr, newVersion);
 
     // 检查是否有 CHANGELOG 更改需要提交
-    const hasChanges = await hasFileChanges('CHANGELOG.md');
+    const hasChanges = await hasChangelogChanges();
     if (hasChanges) {
-      const fullVersion = addVersionPrefix(newVersion);
-      await commitAndPushFile('CHANGELOG.md', COMMIT_TEMPLATES.CHANGELOG_UPDATE(fullVersion), targetBranch);
-      logger.info('✅ CHANGELOG 更新已提交');
+      await commitChangelog(newVersion, targetBranch);
     } else {
       logger.info('CHANGELOG 无更改，跳过提交');
     }
 
     // 🚀 发布到npm - 只对目标分支版本发布
+    const { handleNpmPublish } = await import('./npm');
     await handleNpmPublish(newVersion, targetBranch);
   } catch (error) {
     throw new ActionError(`版本更新和标签创建失败: ${error}`, 'updateVersionAndCreateTag', error);
